rookie/requirements.txt
asgiref==3.3.4
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==3.2.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-mysql==3.9.0
django-redis==4.12.1
djangorestframework==3.12.1
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
jsonpath==0.82
MarkupSafe==1.1.1
mysqlclient==2.0.1
packaging==20.9
parso==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.17
Pygments==2.8.1
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
traitlets==4.3.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
